Align int / long types with Zulip database schema
Fixes #400

+++ src/main/java/com/github/jamesnetherton/zulip/client/api/channel/request/ReorderChannelFoldersApiRequest.java
@@ -22,7 +22,7 @@ public class ReorderChannelFoldersApiRequest extends ZulipApiRequest implements 
      * @param client               The Zulip HTTP client
      * @param channelFolderIdOrder The order of the channel folder ids
      */
-    public ReorderChannelFoldersApiRequest(ZulipHttpClient client, Integer... channelFolderIdOrder) {
+    public ReorderChannelFoldersApiRequest(ZulipHttpClient client, Long... channelFolderIdOrder) {
         super(client);
         putParamAsJsonString(ORDER, channelFolderIdOrder);
     }
